skeleton for both pages
