+++ src/libs/components/ListPageSkeleton/index.ts
@@ -1,0 +1,1 @@
+export * from './ListPageSkeleton';
