+++ src/libs/components/index.ts
@@ -5,3 +5,4 @@ export * from './Header';
 export * from './LanguageChange';
 export * from './Paragraph';
 export * from './DetailsTile';
+export * from './ListPageSkeleton';
