+++ src/modules/Staff/StaffListPage.tsx
@@ -2,76 +2,15 @@ import { useEffect, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import { useGetStaff } from './hooks';
 import { StateType } from 'src/store';
-import { useTranslation } from 'react-i18next';
-import { Header, DetailsTile, Paragraph } from '@UG/libs/components';
+import { Header, DetailsTile, Paragraph, ListPageSkeleton } from '@UG/libs/components';
 import { Academic } from '@UG/libs/types';
 import { styled, useTheme } from '@mui/material/styles';
-import { Box, TextField, Button, Skeleton } from '@mui/material';
-import SearchIcon from '@mui/icons-material/Search';
+import { Box } from '@mui/material';
 import { Link } from 'react-router-dom';
-
-const SearchSkeleton = styled(Skeleton)`
-  width: 710px;
-  height: 60px;
-  border-radius: 25px;
-  margin-right: 20px;
-  margin-bottom: 20px;
-`;
-
-const ButtonSkeleton = styled(Skeleton)`
-  width: 60px;
-  height: 60px;
-  border-radius: 25px;
-`;
-
-const TileSkeleton = styled(Skeleton)`
-  width: 900px;
-  height: 125px;
-  border-radius: 55px;
-  margin-bottom: 40px;
-`;
+import { SearchBar, SearchBarSkeleton } from './SearchBar';
 
 const StyledLink = styled(Link)`
   text-decoration: none;
-`;
-
-const StyledSearchTextField = styled(TextField)`
-  width: 710px;
-  & input {
-    font-family: 'Montserrat';
-    color: ${({ theme }) => theme.palette.primary.dark};
-    margin-left: 20px;
-    margin-right: 20px;
-  }
-  & fieldset {
-    border-radius: 25px;
-  }
-  & .MuiOutlinedInput-root:hover {
-    & fieldset {
-      border: 1px solid ${({ theme }) => theme.palette.primary.main};
-    }
-  }
-  & .MuiFormLabel-root {
-    font-weight: 700;
-    // color: ${({ theme }) => theme.palette.secondary.main};
-  }
-  & label {
-    font-weight: 700;
-    margin-left: 30px;
-    color: ${({ theme }) => theme.palette.secondary.main};
-  }
-  & legend {
-    color: ${({ theme }) => theme.palette.primary.main};
-    margin-left: 30px;
-  }
-`;
-
-const StyledSearchButton = styled(Button)`
-  width: 60px;
-  height: 60px;
-  color: ${({ theme }) => theme.palette.primary.main};
-  border-radius: 25px;
-  font-size: 12px;
 `;
 
 interface StateProps {
@@ -82,7 +21,6 @@ interface StateProps {
 
 export const StaffListPage = () => {
   const theme = useTheme();
-  const { t } = useTranslation();
   const { getStaffList } = useGetStaff();
 
   const { isLoading, staffList, errorMessage } = useSelector<StateType, StateProps>(state => ({
@@ -124,33 +62,8 @@ export const StaffListPage = () => {
     return (
       <>
         <Header />
-        <Box
-          marginLeft="auto"
-          marginRight="auto"
-          marginTop="150px"
-          marginBottom="50px"
-          display="flex"
-          textAlign="center"
-          sx={{ width: 800 }}
-        >
-          <SearchSkeleton animation="wave" variant="rectangular" />
-          <ButtonSkeleton animation="wave" variant="rectangular" />
-        </Box>
-        <Box
-          marginLeft="auto"
-          marginRight="auto"
-          display="flex"
-          alignItems="center"
-          flexDirection="column"
-          sx={{ width: 1080 }}
-        >
-          <TileSkeleton animation="wave" variant="rectangular" />
-          <TileSkeleton animation="wave" variant="rectangular" />
-          <TileSkeleton animation="wave" variant="rectangular" />
-          <TileSkeleton animation="wave" variant="rectangular" />
-          <TileSkeleton animation="wave" variant="rectangular" />
-          <TileSkeleton animation="wave" variant="rectangular" />
-        </Box>
+        <SearchBarSkeleton mb={-10} />
+        <ListPageSkeleton height={100} />
       </>
     );
   }
@@ -158,24 +71,7 @@ export const StaffListPage = () => {
   return (
     <>
       <Header />
-      <Box
-        marginLeft="auto"
-        marginRight="auto"
-        marginTop="150px"
-        component="form"
-        display="flex"
-        justifyContent="space-between"
-        alignItems="center"
-        sx={{
-          width: 800,
-        }}
-        autoComplete="off"
-      >
-        <StyledSearchTextField label={t('search')} data-cy="search-bar" />
-        <StyledSearchButton variant="outlined" data-cy="search-button">
-          <SearchIcon data-cy="search-icon" />
-        </StyledSearchButton>
-      </Box>
+      <SearchBar />
       <Box width={1080} margin="50px auto" display="flex" flexDirection="column" alignItems="center">
         {staffTiles}
       </Box>
